requirements.txt
psutil
fpdf
pdf2docx
pypdf
Pillow
